Factor out date utilities into DateUtils.tsx
